Bed15TrackLine-class {rtracklayer}	R Documentation

Class "Bed15TrackLine"

Description

A UCSC track line for graphical tracks.

Objects from the Class

Objects can be created by calls of the form new("Bed15TrackLine", ...) or parsed from a character vector track line with as(text, "Bed15TrackLine").

Slots

expStep:

A "numeric" scalar indicating the step size for the heatmap color gradient.

expScale:

A positive "numeric" scalar indicating the range of the data to be [-expScale, expScale] for determining the heatmap color gradient.

expNames:

A "character" vector naming the the experimental samples.

name:

Object of class "character" specifying the name of the track.

description:

Object of class "character" describing the track.

visibility:

Object of class "character" indicating the default visible mode of the track, see UCSCTrackModes.

color:

Object of class "integer" representing the track color (as from col2rgb).

priority:

Object of class "numeric" specifying the rank of this track.

Extends

Class "TrackLine", directly.

Methods

as(object, "character")

Export line to its string representation.

Author(s)

Michael Lawrence

References

Official documentation: http://genomewiki.ucsc.edu/index.php/Microarray_track.

See Also

export.bed15 for exporting bed15 tracks.

[Package rtracklayer version 1.50.0 Index]