BigBedSelection-class {rtracklayer}	R Documentation

Selection of ranges and columns

Description

A BigBedSelection represents a query against a BigBed file, see import.bb. It is simply a RangedSelection with colnames parameter.colnames should be a character vector of column names. Default columns are "name", "score", "thick", "itemRgb" and "blocks", if non-empty, as that is the only column supported by BigBed.

Constructor

BigBedSelection(ranges = GRanges(), colnames = "score"): Constructs a BigBedSelection with the given ranges and colnames. a character identifying a genome (see GenomicSelection), or a BigBedFile, in which case the ranges are derived from the bounds of its sequences.

Coercion

as(from, "BigBedSelection"): Coerces from to a BigBedSelection object. Typically, from is a GRanges or a IntegerRangesList, the ranges of which become the ranges in the new BigBedSelection.

Author(s)

Michael Lawrence

Examples

  rl <- IRangesList(chr1 = IRanges::IRanges(c(1, 5), c(3, 6)))

  BigBedSelection(rl)
  as(rl, "BigBedSelection") # same as above

  # do not select any column
  BigBedSelection(rl, character())

[Package rtracklayer version 1.50.0 Index]