IntegerRangesList-methods {rtracklayer}	R Documentation

Ranges on a Genome

Description

Genomic coordinates are often specified in terms of a genome identifier, chromosome name, start position and end position. The rtracklayer package adds convenience methods to IntegerRangesList for the manipulation of genomic ranges. The spaces (or names) of IntegerRangesList are the chromosome names. The universe slot indicates the genome, usually as given by UCSC (e.g. “hg18”).

Accessors

In the code snippets below, x is a IntegerRangesList object.

chrom(x), chrom(x) <- value: Gets or sets the chromosome names for x. This is an alias for names(x).

Author(s)

Michael Lawrence

[Package rtracklayer version 1.50.0 Index]