Quickload-class {rtracklayer}	R Documentation

Quickload Access

Description

The Quickload class represents a Quickload data source, essentially directory layout separating tracks and sequences by genome, along with a few metadata files. This interface abstracts those details and provides access to a Quickload at any URL supported by R (HTTP, FTP, and local files). This is an easy way to make data accessible to the Integrated Genome Browser (IGB).

Constructor

Quickload(uri = "quickload", create = FALSE): Constructs a new Quickload object, representing a repository at uri. If create is TRUE, and uri is writeable (i.e., local), the repository is created if it does not already exist. If it does exist, then a message is emitted to indicate that the repository was not recreated.

Accessor Methods

In the code snippets below, x represents a Quickload object.

x$genome, x[["genome"]]: Get the QuickloadGenome object for the genome named genome. This is where all the data is stored.

length(x): number of genomes in the repository

uri(x): Get the URI pointing to the Quickload repository.

genome(x): Get the identifiers of the genomes present in the repository.

Author(s)

Michael Lawrence

Examples

ql <- Quickload(system.file("tests", "quickload", package = "rtracklayer"))
uri(ql)
genome(ql)
ql$T_species_Oct_2011

[Package rtracklayer version 1.50.0 Index]