TrackDb-class {rtracklayer}	R Documentation

Track Databases

Description

The TrackDb class is an abstraction around a database of tracks. Implementations include BrowserSession derivatives and QuickloadGenome. Here, a track is defined as an interval dataset.

Accessor Methods

Every implementation should support these methods:

length(x): number of tracks

names(x), trackNames(x): names of the tracks

mcols(x): merged metadata on the tracks

track(x, name), x$name, x[[name]]: get the track called name

track(x, name) <- value, x$name <- value, x[[name]] <- value: store the track value under name. Different implementations will support different types for value. Generally, an interval data structure like GenomicRanges.

Author(s)

Michael Lawrence

[Package rtracklayer version 1.50.0 Index]