BasicTrackLine-class {rtracklayer}	R Documentation

Class "BasicTrackLine"

Description

The type of UCSC track line used to annotate most types of tracks (every type except Wiggle).

Objects from the Class

Objects can be created by calls of the form new("BasicTrackLine", ...) or parsed from a character vector track line with as(text, "BasicTrackLine") or converted from a GraphTrackLine using as(wig, "BasicTrackLine").

Slots

itemRgb:

Object of class "logical" indicating whether each feature in a track uploaded as BED should be drawn in its specified color.

useScore:

Object of class "logical" indicating whether the data value should be mapped to color.

group:

Object of class "character" naming a group to which this track should belong.

db:

Object of class "character" indicating the associated genome assembly.

offset:

Object of class "numeric", a number added to all positions in the track.

url:

Object of class "character" referring to additional information about this track.

htmlUrl:

Object of class "character" referring to an HTML page to be displayed with this track.

name:

Object of class "character" specifying the name of the track.

description:

Object of class "character" describing the track.

visibility:

Object of class "character" indicating the default visible mode of the track, see UCSCTrackModes.

color:

Object of class "integer" representing the track color (as from col2rgb).

colorByStrand:

Object of class "matrix" with two columns, as from col2rgb. The two colors indicate the color for each strand (positive, negative).

priority:

Object of class "numeric" specifying the rank of the track.

Extends

Class "TrackLine", directly.

Methods

as(object, "character")

Export line to its string representation.

as(object, "GraphTrackLine")

Convert this line to a graph track line, using defaults for slots not held in common.

Author(s)

Michael Lawrence

References

http://genome.ucsc.edu/goldenPath/help/customTrack.html#TRACK for the official documentation.

See Also

GraphTrackLine for Wiggle/bedGraph tracks.

[Package rtracklayer version 1.50.0 Index]