blocks-methods {rtracklayer}	R Documentation

Get blocks/exons

Description

Obtains the block ranges (subranges, usually exons) from an object, such as a GRanges imported from a BED file.

Usage

blocks(x, ...)

Arguments

x
The instance from which to obtain the block/exon information. Currently must be a GenomicRanges, with a metadata column of name “blocks” and of type IntegerRangesList. Such an object is returned by import.bed and asBED.

...
Additional arguments for methods

Value

A GRangesList with an element for each range in x. The original block ranges are relative to the start of the containing range, so the returned ranges are shifted to absolute coordinates. The seqname and strand are inherited from the containing range.

Author(s)

Michael Lawrence

See Also

import.bed for importing a track from BED, which can store block information; asBED for coercing a GenomicRanges into a BED-like structure that can be passed to this function.

[Package rtracklayer version 1.50.0 Index]