browserSession-methods {rtracklayer}	R Documentation

Get a genome browser session

Description

Methods for getting browser sessions.

Methods

The following methods are defined by rtracklayer.

object = "character"

browserSession(object, ...): Creates a BrowserSession from a genome browser identifier. The identifier corresponds to the prefix of the session class name (e.g. "UCSC" in "UCSCSession"). The arguments in ... are passed to the initialization function of the class.

object = "browserView"

Gets the BrowserSession for the view.

object = "missing"

Calls browserSession("ucsc", ...).

[Package rtracklayer version 1.50.0 Index]