BrowserView-class {rtracklayer}	R Documentation

Class "BrowserView"

Description

An object representing a genome browser view of a particular segment of a genome.

Objects from the Class

A virtual Class: No objects may be created from it directly. See browserView for obtaining an instance of an implementation for a particular genome browser.

Slots

session:

Object of class "BrowserSession" the browser session to which this view belongs.

Methods

This specifies the API implemented by each browser backend. Note that a backend is not guaranteed to support all operations. See the backend-specific documentation for more details. The only built-in backend is UCSCView.

browserSession(object)

Obtains the BrowserSession to which this view belongs.

close(object)

Close this view.

range(object)

Obtains the GRanges displayed by this view.

trackNames(object)

Gets the names of the visible tracks in the view.

trackNames(object) <- value

Sets the visible tracks by their names.

show(object)

Outputs a textual description of this view.

visible(object)

Get a named logical vector indicating whether each track is visible.

visible(object) <- value

Set a logical vector indicating the visibility of each track, with the same names and in the same order as that returned by visible(object).

Author(s)

Michael Lawrence

See Also

browserView for obtaining instances of this class.

[Package rtracklayer version 1.50.0 Index]