liftOver {rtracklayer}	R Documentation

Lift intervals between genome builds

Description

A reimplementation of the UCSC liftover tool for lifting features from one genome build to another. In our preliminary tests, it is significantly faster than the command line tool. Like the UCSC tool, a chain file is required input.

Usage

liftOver(x, chain, ...)

Arguments

x
The intervals to lift-over, usually a GRanges.

chain
A Chain object, usually imported with import.chain, or something coercible to one.

...
Arguments for methods.

Value

A GRangesList object. Each element contains the ranges mapped from the corresponding element in the input (may be one-to-many).

Author(s)

Michael Lawrence

References

http://genome.ucsc.edu/cgi-bin/hgLiftOver

Examples

## Not run: 
chain <- import.chain("hg19ToHg18.over.chain")
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
tx_hg19 <- transcripts(TxDb.Hsapiens.UCSC.hg19.knownGene)
tx_hg18 <- liftOver(tx_hg19, chain)

## End(Not run)

[Package rtracklayer version 1.50.0 Index]