tracks-methods {rtracklayer}	R Documentation

Accessing track names

Description

Methods for getting and setting track names.

Methods

The following methods are defined by rtracklayer for getting track names via the generic trackNames(object, ...).

object = "UCSCSession"

Get the tracks loaded in the session.

object = "UCSCTrackModes"

Get the visible tracks according to the modes (all tracks not set to "hide").

object = "UCSCView"

Get the visible tracks in the view.

The following methods are defined by rtracklayer for setting track names via the generic trackNames(object) <- value.

object = "UCSCTrackModes"

Sets the tracks that should be visible in the modes. All specified tracks with mode "hide" in object are set to mode "full". Any tracks in object that are not specified in the value are set to "hide". No other modes are changed.

object = "UCSCView"

Sets the visible tracks in the view. This opens a new web browser with only the specified tracks visible.

[Package rtracklayer version 1.50.0 Index]