ucscGenomes {rtracklayer}	R Documentation

Get available genomes on UCSC

Description

Get a data.frame describing the available UCSC genomes.

Usage

ucscGenomes(organism=FALSE)

Arguments

organism
A logical(1) indicating whether scientific name should be appended.

Details

For populating the organism column, the web url http://genome.ucsc.edu/cgi-bin is scraped for every assembly version to get the scientific name.

Value

A data.frame with the following columns:

db
UCSC DB identifier (e.g. "hg18")

species
The name of the species (e.g. "Human")

date
The date the genome was built

name
The official name of the genome build

organism
The scientific name of the species (e.g. "Homo sapiens")

Author(s)

Michael Lawrence

See Also

UCSCSession for details on specifying the genome.

Examples

ucscGenomes()

[Package rtracklayer version 1.50.0 Index]