TrackLine-class {rtracklayer}	R Documentation

Class "TrackLine"

Description

An object representing a "track line" in the UCSC format. There are two concrete types of track lines: BasicTrackLine (used for most types of tracks) and GraphTrackLine (used for graphical tracks). This class only declares the common elements between the two.

Objects from the Class

Objects can be created by calls of the form new("TrackLine", ...) or parsed from a character vector track line with as(text, "TrackLine"). But note that UCSC only understands one of the subclasses mentioned above.

Slots

name:

Object of class "character" specifying the name of the track.

description:

Object of class "character" describing the track.

visibility:

Object of class "character" indicating the default visible mode of the track, see UCSCTrackModes.

color:

Object of class "integer" representing the track color (as from col2rgb).

priority:

Object of class "numeric" specifying the rank of this track.

Methods

as(object, "character")

Export line to its string representation.

Author(s)

Michael Lawrence

References

http://genome.ucsc.edu/goldenPath/help/customTrack.html#TRACK for the official documentation.

See Also

BasicTrackLine (used for most types of tracks) and GraphTrackLine (used for Wiggle/bedGraph tracks).

[Package rtracklayer version 1.50.0 Index]