UCSCTrackModes-class {rtracklayer}	R Documentation

Class "UCSCTrackModes"

Description

A vector of view modes ("hide", "dense", "full", "pack", "squish") for each track in a UCSC view.

Objects from the Class

Objects may be created by calls of the form ucscTrackModes(object = character(), hide = character(), dense = character(), pack = character(), squish = character(), full = character()), where object should be a character vector of mode names (with its names attribute specifying the corresponding track names). The other parameters should contain track names that override the modes in object. Later parameters override earlier ones, so, for example, if a track is named in hide and full, it is shown in the full view mode.

Slots

.Data:

Object of class "character" holding the modes ("hide", "dense", "full", "pack", "squish"), with its names attribute holding corresponding track names.

labels:

Object of class "character" holding labels (human-readable names) corresponding to each track/mode.

Extends

Class "character", from data part. Class "vector", by class "character", distance 2.

Methods

trackNames(object)

Gets the names of the visible tracks (those that do not have mode "hide").

trackNames(object) <- value

Sets the names of the visible tracks. Any tracks named in value are set to "full" if the are currently set to "hide" in this object. Any tracks not in value are set to "hide". All other modes are preserved.

object[i]

Gets the track mode of the tracks indexed by i, which can be any type of index supported by character vector subsetting. If i is a character vector, it indexes first by the internal track IDs (the names on .Data) and then by the user-level track names (the labels slot).

object[i] <- value

Sets the track modes indexed by i (in the same way as in object[i] above) to those specified in value.

Author(s)

Michael Lawrence

See Also

UCSCView on which track view modes may be set.

[Package rtracklayer version 1.50.0 Index]