ucscTrackModes-methods {rtracklayer}	R Documentation

Accessing UCSC track modes

Description

Generics for getting and setting UCSC track visibility modes ("hide", "dense", "full", "pack", "squish").

Methods

The following methods are defined by rtracklayer for getting the track modes through the generic ucscTrackModes(object, ...).

object = "character"

function(object, hide = character(), dense = character(), pack = character(), squish = character(), full = character()) Creates an instance of UCSCTrackModes from object, a character vector of mode names, with the corresponding track ids given in the names attribute. Note that object can be a UCSCTrackModes instance, as UCSCTrackModes extends character. The other parameters are character vectors identifying the tracks for each mode and overriding the modes specified by object.

object = "missing"

The same interface as above, except object defaults to an empty character vector.

object = "UCSCView"

Gets modes for tracks in the view.

object = "UCSCSession"

Gets default modes for the tracks in the session. These are the modes that will be used as the default for a newly created view.

The following methods are defined by rtracklayer for setting the track modes through the generic ucscTrackModes(object) <- value.

object = "UCSCView", value = "UCSCTrackModes"

Sets the modes for the tracks in the view.

object = "UCSCView", value = "character"

Sets the modes from a character vector of mode names, with the corresponding track names given in the names attribute.

See Also

trackNames and trackNames<- for just getting or setting which tracks are visible (not of mode "hide").

Examples

# Tracks "foo" and "bar" are fully shown, "baz" is hidden
modes <- ucscTrackModes(full = c("foo", "bar"), hide = "baz")
# Update the modes to hide track "bar"
modes2 <- ucscTrackModes(modes, hide = "bar")

[Package rtracklayer version 1.50.0 Index]