wigToBigWig {rtracklayer}	R Documentation

Convert WIG to BigWig

Description

This function calls the Kent C library to efficiently convert a WIG file to a BigWig file, without loading the entire file into memory. This solves the problem where simple tools write out text WIG files, instead of more efficiently accessed binary BigWig files.

Usage

wigToBigWig(x, seqinfo,
            dest = paste(file_path_sans_ext(x, TRUE), "bw", sep = "."),
            clip = FALSE)

Arguments

x
Path or URL to the WIG file. Connections are not supported.

seqinfo
Seqinfo object, describing the genome of the data. All BigWig files must have this defined.

dest
The path to which to write the BigWig file. Defaults to x with the extension changed to “bw”.

clip
If TRUE, regions outside of seqinfo will be clipped, so that no error is thrown.

Author(s)

Michael Lawrence

See Also

BigWig import and export support

[Package rtracklayer version 1.50.0 Index]