GraphTrackLine-class {rtracklayer}	R Documentation

Class "GraphTrackLine"

Description

A UCSC track line for graphical tracks.

Objects from the Class

Objects can be created by calls of the form new("GraphTrackLine", ...) or parsed from a character vector track line with as(text, "GraphTrackLine") or converted from a BasicTrackLine using as(basic, "GraphTrackLine").

Slots

altColor:

Object of class "integer" giving an alternate color, as from col2rgb.

autoScale:

Object of class "logical" indicating whether to automatically scale to min/max of the data.

alwaysZero:

Object of class "logical" indicating whether to fix the lower limit of the Y axis at zero.

gridDefault:

Object of class "logical" indicating whether a grid should be drawn.

maxHeightPixels:

Object of class "numeric" of length three (max, default, min), giving the allowable range for the vertical height of the graph.

graphType:

Object of class "character", specifying the graph type, either "bar" or "points".

viewLimits:

Object of class "numeric" and of length two specifying the data range (min, max) shown in the graph.

yLineMark:

Object of class "numeric" giving the position of a horizontal line.

yLineOnOff:

Object of class "logical" indicating whether the yLineMark should be visible.

windowingFunction:

Object of class "character", one of "maximum", "mean", "minimum", for removing points when the graph shrinks.

smoothingWindow:

Object of class "numeric" giving the window size of a smoother to pass over the graph.

type:

Scalar "character" indicating the type of the track, either "wig" or "bedGraph".

name:

Object of class "character" specifying the name of the track.

description:

Object of class "character" describing the track.

visibility:

Object of class "character" indicating the default visible mode of the track, see UCSCTrackModes.

color:

Object of class "integer" representing the track color (as from col2rgb).

priority:

Object of class "numeric" specifying the rank of this track.

Extends

Class "TrackLine", directly.

Methods

as(object, "character")

Export line to its string representation.

as(object, "BasicTrackLine")

Convert this line to a basic UCSC track line, using defaults for slots not held in common.

Author(s)

Michael Lawrence

References

Official documentation: http://genome.ucsc.edu/goldenPath/help/wiggle.html.

See Also

export.wig, export.bedGraph for exporting graphical tracks.

[Package rtracklayer version 1.50.0 Index]