browserView-methods {rtracklayer}	R Documentation

Getting browser views

Description

Methods for creating and getting browser views.

Usage

  browserView(object, range, track, ...)

Arguments

object
The object from which to get the views.

range
The GRanges or IntegerRangesList to display. If there are multiple elements, a view is created for each element and a BrowserViewList is returned.

track
List of track names to make visible in the view.

...
Arguments to pass to methods

Methods

The following methods are defined by rtracklayer.

object = "UCSCSession"

browserView(object, range = range(object), track = trackNames(object), imagewidth = 800, ...): Creates a BrowserView of range with visible tracks specified by track. The imagewidth parameter specifies the width of the track image in pixels. track may be an instance of UCSCTrackModes. Arguments in ... are passed to ucscTrackModes to create the UCSCTrackModes instance that will override modes indicated by the track parameter.

Examples

## Not run: 
  session <- browserSession()
  browserView(session,
              GRangesForUCSCGenome("hg19", "chr2", IRanges(20000, 50000)))
  ## only view "knownGene" track
  browserView(session, track = "knownGene")

## End(Not run)

[Package rtracklayer version 1.50.0 Index]