UCSCView-class {rtracklayer}	R Documentation

Class "UCSCView"

Description

An object representing a view of a genome in the UCSC browser.

Objects from the Class

Calling browserView(session, range = range(object), track = trackNames(object), browse = TRUE, ...) creates BrowserView of range with visible tracks specified by track. track may be an instance of UCSCTrackModes. Arguments in ... should match parameters to a ucscTrackModes method for creating a UCSCTrackModes instance that will be merged with and override modes indicated by the track parameter. If browse is TRUE (default), automatically launch a browser to display the view.

Slots

hgsid:

Object of class "numeric", which identifies this view to UCSC.

session:

Object of class "BrowserSession" to which this view belongs.

Extends

Class "BrowserView", directly.

Methods

activeView(object)

Obtains a logical indicating whether this view is the active view.

range(object)

Obtains the GRanges displayed by this view.

range(object) <- value

Sets the GRanges or IntegerRangesList displayed by this view.

trackNames(object)

Gets the names of the visible tracks in this view.

trackNames(object) <- value

Sets the visible tracks by name.

visible(object)

Get a named logical vector indicating whether each track is visible.

visible(object) <- value

Set a logical vector indicating the visibility of each track, in the same order as returned by visible(object).

ucscTrackModes(object)

Obtains the UCSCTrackModes for this view.

ucscTrackModes(object) <- value

Sets the UCSCTrackModes for this view. The value may be either a UCSCTrackModes instance or a character vector that will be coerced by a call to ucscTrackModes.

Utilities

viewURL(x)

Gets the URL corresponding to the view x. Typically used when passing browse=FALSE to the constructor in order to display the view in another way.

Author(s)

Michael Lawrence

See Also

browserView for creating instances of this class.

[Package rtracklayer version 1.50.0 Index]